rdmaxcel_sys

Function ibv_create_ah_from_wc

pub unsafe extern "C" fn ibv_create_ah_from_wc(
    pd: *mut ibv_pd,
    wc: *mut ibv_wc,
    grh: *mut ibv_grh,
    port_num: u8,
) -> *mut ibv_ah

ibv_create_ah_from_wc - Creates an address handle associated with the sender of the specified work completion. @pd: The protection domain associated with the address handle. @wc: Work completion information associated with a received message. @grh: References the received global route header. This parameter is ignored unless the work completion indicates that the GRH is valid. @port_num: The outbound port number to associate with the address.

The address handle is used to reference a local or global destination in all UD QP post sends.