rdmaxcel_sys

Type Alias CUmemAllocationProp

pub type CUmemAllocationProp = CUmemAllocationProp_v1;

Specifies the allocation properties for a allocation.

Aliased Type

#[repr(C)]
pub struct CUmemAllocationProp {
    pub type_: u32,
    pub requestedHandleTypes: u32,
    pub location: CUmemLocation_st,
    pub win32HandleMetaData: *mut c_void,
    pub allocFlags: CUmemAllocationProp_st__bindgen_ty_1,
}

Fields
type_: u32

Allocation type

requestedHandleTypes: u32

requested ::CUmemAllocationHandleType

location: CUmemLocation_st

Location of allocation

win32HandleMetaData: *mut c_void

Windows-specific POBJECT_ATTRIBUTES required when ::CU_MEM_HANDLE_TYPE_WIN32 is specified. This object attributes structure includes security attributes that define the scope of which exported allocations may be transferred to other processes. In all other cases, this field is required to be zero.

allocFlags: CUmemAllocationProp_st__bindgen_ty_1